### 2024-01-31 #6308 @editorialbot
[PRE REVIEW]: Pycausal-Explorer: A Scikit-learn compatible causal inference toolkit
**Submitting author:** @gotolino (<a href="http://orcid.org/0009-0000-2451-7029">Guilherme Goto Escudero</a>)
**Repository:** https://github.com/gotolino/pycausal-explorer
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
