### 2024-03-20 #6308 @srmnitc
> 👋 @srmnitc - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

This submission looks interesting, but I am currently editing for JOSS. So I have to unfortunately skip this one.
