### 2024-04-15 #6549 @saeedahmadicp
Dear @teonbrooks, I have carefully reviewed the submission you sent. The article and the accompanying package aim to provide a Python module for causal inference and treatment effect estimation. While the Package is well-documented and contains example use cases and automated tests, as well as a well-written paper with all the necessary sections, there are a few issues that make the current version of the paper unsuitable for publication in the JOSS journal.

Firstly, the contribution of the package lacks novelty and sufficient need as there are already many similar packages available, such as Casual ML by Uber, which is open-sourced, regularly maintained, and widely used. Additionally, several other packages are available from Microsoft. The authors have failed to highlight any unique features of their package, making the scholarly impact of the paper insufficient.

Secondly, the authors haven't provided any guidelines for contributions to the repository. Additionally, there seems to be a conflict between the authors of the paper, as the fourth author in the GitHub contribution is different from the paper's authors i.e.,  Roseli de Deus Lopes.

[ ${\color{blue}Note}$ This is my first review in this journal, so let me know if I have missed anything @teonbrooks. ]
