### 2024-11-26 #6549 @sara-02
@crvernon I have reviewed the package. I was able to install it from pip and run the examples provided in their Readme as well as in the documentation without any error.

I second the points raised by @saeedahmadicp. Overall, the primary concerns are:
1. The paper needs contributor guidelines.
2. Because of point 1, it is not clear how to run the tests. While I was still able to run the tests via pytest, and the tests passed, new users and contributors may need help navigating it.
3. The paper does not provide a functionality or performance comparison with the software mentioned in the paper from Uber and Microsoft.

Writing issue:
Lines 35-36 in the paper pdf appear broken abruptly; it is a spacing issue.

Minor suggestion:
The example provided in the document requires users to download a new dataset, which takes away from the flow of reading. It could be better to give an example which uses one of the two preloaded datasets that are already a part of the library.
